- add company filter to lead segment - add tests

+++ Segment/Query/Filter/SegmentReferenceFilterQueryBuilder.php
@@ -54,6 +54,15 @@ class SegmentReferenceFilterQueryBuilder extends BaseFilterQueryBuilder implemen
             throw new \RuntimeException('The supported field is '.CompanySegmentModel::PROPERTIES_FIELD);
         }
 
+        if (!array_key_exists(MAUTIC_TABLE_PREFIX.'companies', $queryBuilder->getTableAliases())) {
+            return $this->applyQueryToLeadSegment($queryBuilder, $filter);
+        }
+
+        return $this->applyQueryToCompanySegment($queryBuilder, $filter);
+    }
+
+    private function applyQueryToCompanySegment(QueryBuilder $queryBuilder, ContactSegmentFilter $filter): QueryBuilder
+    {
         $companiesTableAlias = $queryBuilder->getTableAlias(MAUTIC_TABLE_PREFIX.'companies');
         \assert(is_string($companiesTableAlias));
         $segmentIds = $filter->getParameterValue();
@@ -116,6 +125,100 @@ class SegmentReferenceFilterQueryBuilder extends BaseFilterQueryBuilder implemen
                 $queryBuilder->addLogic($expression, $filter->getGlue());
             }
 
+            // Preserve memory and detach segments that are not needed anymore.
+            $this->entityManager->detach($companySegment);
+        }
+
+        if (count($orLogic) > 0) {
+            $queryBuilder->addLogic(new CompositeExpression(CompositeExpression::TYPE_OR, $orLogic), $filter->getGlue());
+        }
+
+        return $queryBuilder;
+    }
+
+    private function applyQueryToLeadSegment(QueryBuilder $queryBuilder, ContactSegmentFilter $filter): QueryBuilder
+    {
+        $leadAlias               = $queryBuilder->getTableAlias(MAUTIC_TABLE_PREFIX.'leads');
+        $companiesLeadTableAlias = $this->generateRandomParameterName();
+        $queryBuilder->leftJoin(
+            $leadAlias,
+            MAUTIC_TABLE_PREFIX.'companies_leads',
+            $companiesLeadTableAlias,
+            $companiesLeadTableAlias.'.lead_id = '.$leadAlias.'.id AND '.$companiesLeadTableAlias.'.is_primary = 1'
+        );
+        $companiesTableAlias = $queryBuilder->getTableAlias(MAUTIC_TABLE_PREFIX.'companies');
+        if (empty($companiesTableAlias)) {
+            $companiesTableAlias = 'bbbboo';
+        }
+
+        \assert(is_string($companiesTableAlias));
+
+        $segmentIds = $filter->getParameterValue();
+        if (empty($segmentIds)) {
+            return $queryBuilder;
+        }
+        \assert(is_array($segmentIds) || is_numeric($segmentIds));
+
+        if (!is_array($segmentIds)) {
+            $segmentIds = [(int) $segmentIds];
+        }
+
+        $orLogic           = [];
+        foreach ($segmentIds as $segmentId) {
+            $exclusion = in_array($filter->getOperator(), ['notExists', 'notIn'], true);
+
+            /** @var CompanySegment|null $companySegment */
+            $companySegment    = $this->entityManager->getRepository(CompanySegment::class)->find($segmentId);
+
+            if (null === $companySegment) {
+                throw new SegmentNotFoundException(sprintf('Segment %d used in the filter does not exist anymore.', $segmentId));
+            }
+
+            $contactSegment = new CompanySegmentAsLeadSegment($companySegment);
+            $filters        = $this->leadSegmentFilterFactory->getSegmentFilters($contactSegment);
+            $segmentQueryBuilder = $this->companySegmentQueryBuilder->assembleCompaniesSegmentQueryBuilderLeadSegment(
+                $companySegment,
+                $filters,
+                true
+            );
+            $subSegmentCompaniesTableAlias = $segmentQueryBuilder->getTableAlias(MAUTIC_TABLE_PREFIX.'companies') ?? $this->generateRandomParameterName();
+
+            \assert(is_string($subSegmentCompaniesTableAlias));
+            $segmentQueryBuilder->resetQueryParts(['select'])->select('null');
+
+            // If the segment contains no filters; it means its for manually subscribed only
+            if (count($filters) > 0) {
+                //                dump($companySegment->getId());
+                $segmentQueryBuilder = $this->companySegmentQueryBuilder->addManuallyUnsubscribedQuery($segmentQueryBuilder, $companySegment);
+            }
+
+            $segmentQueryBuilder = $this->companySegmentQueryBuilder->addManuallySubscribedQuery($segmentQueryBuilder, $companySegment);
+            // This query looks a bit too complex, but if the segment(s) has more or less complex filter this is (probably)
+            // the way to go. Hours spent optimizing: 3. Increment if you spent yet more here.
+            $segmentQueryBuilder = $this->companySegmentQueryBuilder->addCompanySegmentQuery($segmentQueryBuilder, $companySegment);
+
+            $parameters = $segmentQueryBuilder->getParameters();
+            foreach ($parameters as $key => $value) {
+                $queryBuilder->setParameter($key, $value);
+            }
+
+            $this->companySegmentQueryBuilder->queryBuilderGenerated($companySegment, $segmentQueryBuilder);
+
+            $segmentQueryWherePart = $segmentQueryBuilder->getQueryPart('where');
+
+            $segmentQueryBuilder->where(sprintf('%s.company_id = %s.id', $companiesLeadTableAlias, $subSegmentCompaniesTableAlias));
+            $segmentQueryBuilder->andWhere($segmentQueryWherePart);
+            if ($exclusion) {
+                $expression = $queryBuilder->expr()->notExists($segmentQueryBuilder->getSQL());
+            } else {
+                $expression = $queryBuilder->expr()->exists($segmentQueryBuilder->getSQL());
+            }
+
+            if (!$exclusion && count($segmentIds) > 1) {
+                $orLogic[] = $expression;
+            } else {
+                $queryBuilder->addLogic($expression, $filter->getGlue());
+            }
             // Preserve memory and detach segments that are not needed anymore.
             $this->entityManager->detach($companySegment);
         }
